refactor/#11: Unify Member entity
