--- src/main/java/com/picpic/server/room/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.picpic.server.room.repository;
-
-import com.picpic.server.room.entity.Member;
-import jdk.jfr.Registered;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
